Moved authentication route to separate file, tweaked docs to reflect new authentication requirements.
