+++ api/docs/api_data.js
@@ -1,1 +1,1 @@
-define({ "api": [  {    "type": "post",    "url": "api/geocaches",    "title": "Creates a new geocache",    "name": "CreateGeocache",    "group": "Geocaches",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "name",            "description": "<p>Name of the new geocache</p>"          },          {            "group": "Parameter",            "type": "String",            "optional": true,            "field": "description",            "description": "<p>Description of the new geocache (optional)</p>"          },          {            "group": "Parameter",            "type": "Number",            "optional": false,            "field": "createdBy",            "description": "<p>User id of the new geocache's creator</p>"          },          {            "group": "Parameter",            "type": "Number",            "optional": false,            "field": "xCoord",            "description": "<p>X-coordinate of the new geocache</p>"          },          {            "group": "Parameter",            "type": "Number",            "optional": false,            "field": "yCoord",            "description": "<p>Y-coordinate of the new geocache</p>"          }        ]      }    },    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"_id\": \"57314883c1d0de878fb4f137\",\n  \"yCoord\": 157.312452463,\n  \"xCoord\": -131.31785312,\n  \"createdBy\": 234567h8j90o987654,\n  \"description\": \"Test Description\",\n  \"name\": \"John's Cache\",\n  \"__v\": 0,\n  \"logs\": [],\n  \"updated\": \"2016-05-10T02:33:39.129Z\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "routes/geocaches.js",    "groupTitle": "Geocaches"  },  {    "type": "delete",    "url": "api/geocaches/{geocache_id}",    "title": "Deletes a geocache by id",    "name": "DeleteGeocache",    "group": "Geocaches",    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"message\": \"Geocache successfully deleted\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "routes/geocaches.js",    "groupTitle": "Geocaches"  },  {    "type": "get",    "url": "api/geocaches/{geocache_id}",    "title": "Requests geocache information",    "name": "GetGeocacheFromId",    "group": "Geocaches",    "success": {      "fields": {        "Success 200": [          {            "group": "Success 200",            "type": "String",            "optional": false,            "field": "name",            "description": "<p>Name of the geocache</p>"          },          {            "group": "Success 200",            "type": "String",            "optional": false,            "field": "description",            "description": "<p>Description of the geocache</p>"          },          {            "group": "Success 200",            "type": "User",            "optional": false,            "field": "createdBy",            "description": "<p>User information of the geocache's creator</p>"          },          {            "group": "Success 200",            "type": "Number",            "optional": false,            "field": "xCoord",            "description": "<p>X-coordinate of the geocache</p>"          },          {            "group": "Success 200",            "type": "Number",            "optional": false,            "field": "yCoord",            "description": "<p>Y-coordinate of the geocache</p>"          },          {            "group": "Success 200",            "type": "Log[]",            "optional": false,            "field": "logs",            "description": "<p>Array of log information</p>"          },          {            "group": "Success 200",            "type": "Date",            "optional": false,            "field": "updated",            "description": "<p>Date geocache was last updated</p>"          }        ]      },      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n    \"_id\": \"57314883c1d0de878fb4f137\",\n    \"yCoord\": 157.312452463,\n    \"xCoord\": -131.31785312,\n    \"createdBy\": {\n        \"id\": \"573147edc1d0de878fb4f136\",\n        \"username\": \"JohnDoe\",\n        \"email\": \"john@examplemail.com\",\n        \"memberSince\": \"2016-05-10T02:31:09.692Z\",\n        \"geocachesVisited\": [],\n        \"geocachesCreated\": [\n            \"57314883c1d0de878fb4f137\"\n        ]\n    },\n    \"description\": \"Test Description\",\n    \"name\": \"John's Cache\",\n    \"__v\": 0,\n    \"logs\": [],\n    \"updated\": \"2016-05-10T02:33:39.129Z\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "routes/geocaches.js",    "groupTitle": "Geocaches"  },  {    "type": "get",    "url": "api/geocaches",    "title": "Requests all geocache information",    "name": "GetGeocaches",    "group": "Geocaches",    "success": {      "fields": {        "Success 200": [          {            "group": "Success 200",            "type": "String",            "optional": false,            "field": "name",            "description": "<p>Name of the geocache</p>"          },          {            "group": "Success 200",            "type": "String",            "optional": false,            "field": "description",            "description": "<p>Description of the geocache</p>"          },          {            "group": "Success 200",            "type": "User",            "optional": false,            "field": "createdBy",            "description": "<p>User information of the geocache's creator</p>"          },          {            "group": "Success 200",            "type": "Number",            "optional": false,            "field": "xCoord",            "description": "<p>X-coordinate of the geocache</p>"          },          {            "group": "Success 200",            "type": "Number",            "optional": false,            "field": "yCoord",            "description": "<p>Y-coordinate of the geocache</p>"          },          {            "group": "Success 200",            "type": "Log[]",            "optional": false,            "field": "logs",            "description": "<p>Array of log information</p>"          },          {            "group": "Success 200",            "type": "Date",            "optional": false,            "field": "updated",            "description": "<p>Date geocache was last updated</p>"          }        ]      },      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n[\n {\n     \"_id\": \"57314883c1d0de878fb4f137\",\n     \"yCoord\": 157.312452463,\n     \"xCoord\": -131.31785312,\n     \"createdBy\": {\n         \"id\": \"573147edc1d0de878fb4f136\",\n         \"username\": \"JohnDoe\",\n         \"email\": \"john@examplemail.com\",\n         \"memberSince\": \"2016-05-10T02:31:09.692Z\",\n         \"geocachesVisited\": [],\n         \"geocachesCreated\": [\n             \"57314883c1d0de878fb4f137\"\n         ]\n     },\n     \"description\": \"Test Description\",\n     \"name\": \"John's Cache\",\n     \"__v\": 0,\n     \"logs\": [],\n     \"updated\": \"2016-05-10T02:33:39.129Z\"\n },\n {\n     \"_id\": \"57314883c1d0de878fb4f138\",\n     \"yCoord\": 157.234452463,\n     \"xCoord\": -130.317321412,\n     \"createdBy\": {\n         \"id\": \"573147edc1d0de878fb4f135\",\n         \"username\": \"JaneDoe\",\n         \"email\": \"jane@examplemail.com\",\n         \"memberSince\": \"2015-12-11T02:31:09.692Z\",\n         \"geocachesVisited\": [],\n         \"geocachesCreated\": [\n             \"57314883c1d0de878fb4f138\"\n         ]\n     },\n     \"description\": \"Test Description 2\",\n     \"name\": \"Jane's Cache\",\n     \"__v\": 0,\n     \"logs\": [],\n     \"updated\": \"2016-05-10T02:33:39.129Z\"\n }\n]",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "routes/geocaches.js",    "groupTitle": "Geocaches"  },  {    "type": "put",    "url": "api/geocaches/{geocache_id}",    "title": "Updates a geocache",    "name": "UpdateGeocache",    "group": "Geocaches",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "name",            "description": "<p>New name of the geocache</p>"          },          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "description",            "description": "<p>New description of the geocache</p>"          },          {            "group": "Parameter",            "type": "Number",            "optional": false,            "field": "xCoord",            "description": "<p>New X-coordinate of the geocache</p>"          },          {            "group": "Parameter",            "type": "Number",            "optional": false,            "field": "yCoord",            "description": "<p>New Y-coordinate of the geocache</p>"          }        ]      }    },    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"message\": \"Geocache has been updated\",\n  \"user\": {\n    \"_id\": \"57314883c1d0de878fb4f137\",\n    \"yCoord\": 107.312452463,\n    \"xCoord\": -121.31785312,\n    \"createdBy\": {\n        \"id\": \"573147edc1d0de878fb4f136\",\n        \"username\": \"JohnDoe\",\n        \"email\": \"john@examplemail.com\",\n        \"memberSince\": \"2016-05-10T02:31:09.692Z\",\n        \"geocachesVisited\": [],\n        \"geocachesCreated\": [\n            \"57314883c1d0de878fb4f137\"\n        ]\n    },\n    \"description\": \"Trial Description\",\n    \"name\": \"John's Better Cache\",\n    \"__v\": 0,\n    \"logs\": [],\n    \"updated\": \"2016-05-10T02:33:39.129Z\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "routes/geocaches.js",    "groupTitle": "Geocaches"  },  {    "type": "post",    "url": "api/users",    "title": "Creates a new user",    "name": "CreateUser",    "group": "Users",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "username",            "description": "<p>Username of the new user</p>"          },          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "password",            "description": "<p>Password of the new user</p>"          },          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "email",            "description": "<p>Email of the new user</p>"          }        ]      }    },    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"message\": \"User was created\",\n  \"user\": {\n     \"_id\": \"1q2w3e4r56t7yu8\"\n     \"username\": \"JohnDoe\",\n     \"email\": \"john@examplemail.com\"\n     \"geocachesVisited\": \"[1234wer45rty73e2d, q2w3e4df5t6y879u]\"\n     \"geocachesCreated\": \"[q2w3e4df5t6y879u]\"\n  }\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "routes/users.js",    "groupTitle": "Users"  },  {    "type": "delete",    "url": "api/users/{user_id}",    "title": "Deletes a user by id",    "name": "DeleteUser",    "group": "Users",    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"message\": \"User successfully deleted\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "routes/users.js",    "groupTitle": "Users"  },  {    "type": "get",    "url": "api/users/{user_id}",    "title": "Requests user information",    "name": "GetUserFromId",    "group": "Users",    "success": {      "fields": {        "Success 200": [          {            "group": "Success 200",            "type": "String",            "optional": false,            "field": "_id",            "description": "<p>Unique id of the user</p>"          },          {            "group": "Success 200",            "type": "String",            "optional": false,            "field": "username",            "description": "<p>Username of the user</p>"          },          {            "group": "Success 200",            "type": "String",            "optional": false,            "field": "email",            "description": "<p>Email of the user</p>"          },          {            "group": "Success 200",            "type": "Date",            "optional": false,            "field": "memberSince",            "description": "<p>Date when user joined</p>"          },          {            "group": "Success 200",            "type": "Geocache[]",            "optional": false,            "field": "geocachesVisited",            "description": "<p>Geocaches that the user has logged visits</p>"          },          {            "group": "Success 200",            "type": "Geocache[]",            "optional": false,            "field": "geocachesCreated",            "description": "<p>Geocaches that the user has created</p>"          }        ]      },      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n     \"_id\": \"1q2w3e4r56t7yu8\"\n     \"username\": \"JohnDoe\",\n     \"email\": \"john@examplemail.com\"\n     \"memberSince\": \"2016-05-10T02:31:09.692Z\"\n     \"geocachesVisited\": \"[]\"\n     \"geocachesCreated\": \"[]\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "routes/users.js",    "groupTitle": "Users"  },  {    "type": "get",    "url": "api/users",    "title": "Requests all user information",    "name": "GetUsers",    "group": "Users",    "success": {      "fields": {        "Success 200": [          {            "group": "Success 200",            "type": "String",            "optional": false,            "field": "_id",            "description": "<p>Unique id of the user</p>"          },          {            "group": "Success 200",            "type": "String",            "optional": false,            "field": "username",            "description": "<p>Username of the user</p>"          },          {            "group": "Success 200",            "type": "String",            "optional": false,            "field": "email",            "description": "<p>Email of the user</p>"          },          {            "group": "Success 200",            "type": "Date",            "optional": false,            "field": "memberSince",            "description": "<p>Date when user joined</p>"          },          {            "group": "Success 200",            "type": "Geocache[]",            "optional": false,            "field": "geocachesVisited",            "description": "<p>Geocaches that the user has logged visits</p>"          },          {            "group": "Success 200",            "type": "Geocache[]",            "optional": false,            "field": "geocachesCreated",            "description": "<p>Geocaches that the user has created</p>"          }        ]      },      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n[\n {\n     \"_id\": \"1q2w3e4r56t7yu8\"\n     \"username\": \"JohnDoe\",\n     \"email\": \"john@examplemail.com\"\n     \"memberSince\": \"2016-05-10T02:31:09.692Z\"\n     \"geocachesVisited\": \"[]\"\n     \"geocachesCreated\": \"[]\"\n },\n {\n     \"_id\": \"9u89u7t8eyqwqwdq0\"\n     \"username\": \"JaneDoe\",\n     \"email\": \"jane@examplemail.com\"\n     \"memberSince\": \"\"2016-05-10T01:31:19.692Z\"\"\n     \"geocachesVisited\": \"[]\"\n     \"geocachesCreated\": \"[]\"\n }\n]",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "routes/users.js",    "groupTitle": "Users"  },  {    "type": "put",    "url": "api/users/{user_id}",    "title": "Updates a user",    "name": "UpdateUser",    "group": "Users",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "username",            "description": "<p>New username of the user</p>"          },          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "email",            "description": "<p>New email of the user</p>"          }        ]      }    },    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"message\": \"User has been updated\",\n  \"user\": {\n     \"_id\": \"1q2w3e4r56t7yu8\"\n     \"username\": \"JohnDoel\",\n     \"email\": \"johnnyboy@examplemail.com\"\n     \"geocachesVisited\": \"[1234wer45rty73e2d, q2w3e4df5t6y879u]\"\n     \"geocachesCreated\": \"[q2w3e4df5t6y879u]\"\n  }\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "routes/users.js",    "groupTitle": "Users"  }] });
+define({ "api": [  {    "type": "post",    "url": "api/authentication",    "title": "Authenticates an existing user",    "name": "AuthenticateUser",    "group": "Authentication",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "username",            "description": "<p>Username of the user</p>"          },          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "password",            "description": "<p>Password of the user</p>"          }        ]      }    },    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"success\": \"true\",\n  \"message\": \"Enjoy your token!\",\n  \"token\": \"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyIkX18iOnsic3RyaWN0T\n  W9kZSI6dHJ1ZSwiZ2V0dGVycyI6e30sIndhc1BvcHVsYXRlZCI6ZmFsc2UsImFjdGl2Z\n  VBhdGhzIjp7InBhdGhzIjp7ImVtYWlsIjoiaW5pdCIsInBhc3N3b3JkIjoiaW5pdCIsI\n  nVzZXJuYW1lIjoiaW5pdCIsIm1lbWJlclNpbmNlIjoiaW5pdCIsImdlb2NhY2hlc1Zpc\n  2l0ZWQiOiJpbml0IiwiZ2VvY2FjaGVzQ3JlYXRlZCI6ImluaXQiLCJfX3YiOiJpbml0I\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "routes/authentication.js",    "groupTitle": "Authentication"  },  {    "type": "post",    "url": "api/geocaches",    "title": "Creates a new geocache",    "name": "CreateGeocache",    "group": "Geocaches",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "name",            "description": "<p>Name of the new geocache</p>"          },          {            "group": "Parameter",            "type": "String",            "optional": true,            "field": "description",            "description": "<p>Description of the new geocache (optional)</p>"          },          {            "group": "Parameter",            "type": "Number",            "optional": false,            "field": "createdBy",            "description": "<p>User id of the new geocache's creator</p>"          },          {            "group": "Parameter",            "type": "Number",            "optional": false,            "field": "xCoord",            "description": "<p>X-coordinate of the new geocache</p>"          },          {            "group": "Parameter",            "type": "Number",            "optional": false,            "field": "yCoord",            "description": "<p>Y-coordinate of the new geocache</p>"          },          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "token",            "description": "<p>Authentication token of existing user (see api/authentication).</p>"          }        ]      }    },    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"_id\": \"57314883c1d0de878fb4f137\",\n  \"yCoord\": 157.312452463,\n  \"xCoord\": -131.31785312,\n  \"createdBy\": 234567h8j90o987654,\n  \"description\": \"Test Description\",\n  \"name\": \"John's Cache\",\n  \"__v\": 0,\n  \"logs\": [],\n  \"updated\": \"2016-05-10T02:33:39.129Z\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "routes/geocaches.js",    "groupTitle": "Geocaches"  },  {    "type": "delete",    "url": "api/geocaches/{geocache_id}",    "title": "Deletes a geocache by id",    "name": "DeleteGeocache",    "group": "Geocaches",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "token",            "description": "<p>Authentication token of existing user (see api/authentication).</p>"          }        ]      }    },    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"message\": \"Geocache successfully deleted\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "routes/geocaches.js",    "groupTitle": "Geocaches"  },  {    "type": "get",    "url": "api/geocaches/{geocache_id}",    "title": "Requests geocache information",    "name": "GetGeocacheFromId",    "group": "Geocaches",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "token",            "description": "<p>Authentication token of existing user (see api/authentication).</p>"          }        ]      }    },    "success": {      "fields": {        "Success 200": [          {            "group": "Success 200",            "type": "String",            "optional": false,            "field": "name",            "description": "<p>Name of the geocache</p>"          },          {            "group": "Success 200",            "type": "String",            "optional": false,            "field": "description",            "description": "<p>Description of the geocache</p>"          },          {            "group": "Success 200",            "type": "User",            "optional": false,            "field": "createdBy",            "description": "<p>User information of the geocache's creator</p>"          },          {            "group": "Success 200",            "type": "Number",            "optional": false,            "field": "xCoord",            "description": "<p>X-coordinate of the geocache</p>"          },          {            "group": "Success 200",            "type": "Number",            "optional": false,            "field": "yCoord",            "description": "<p>Y-coordinate of the geocache</p>"          },          {            "group": "Success 200",            "type": "Log[]",            "optional": false,            "field": "logs",            "description": "<p>Array of log information</p>"          },          {            "group": "Success 200",            "type": "Date",            "optional": false,            "field": "updated",            "description": "<p>Date geocache was last updated</p>"          }        ]      },      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n    \"_id\": \"57314883c1d0de878fb4f137\",\n    \"yCoord\": 157.312452463,\n    \"xCoord\": -131.31785312,\n    \"createdBy\": {\n        \"id\": \"573147edc1d0de878fb4f136\",\n        \"username\": \"JohnDoe\",\n        \"email\": \"john@examplemail.com\",\n        \"memberSince\": \"2016-05-10T02:31:09.692Z\",\n        \"geocachesVisited\": [],\n        \"geocachesCreated\": [\n            \"57314883c1d0de878fb4f137\"\n        ]\n    },\n    \"description\": \"Test Description\",\n    \"name\": \"John's Cache\",\n    \"__v\": 0,\n    \"logs\": [],\n    \"updated\": \"2016-05-10T02:33:39.129Z\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "routes/geocaches.js",    "groupTitle": "Geocaches"  },  {    "type": "get",    "url": "api/geocaches",    "title": "Requests all geocache information",    "name": "GetGeocaches",    "group": "Geocaches",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "token",            "description": "<p>Authentication token of existing user (see api/authentication).</p>"          }        ]      }    },    "success": {      "fields": {        "Success 200": [          {            "group": "Success 200",            "type": "String",            "optional": false,            "field": "name",            "description": "<p>Name of the geocache</p>"          },          {            "group": "Success 200",            "type": "String",            "optional": false,            "field": "description",            "description": "<p>Description of the geocache</p>"          },          {            "group": "Success 200",            "type": "User",            "optional": false,            "field": "createdBy",            "description": "<p>User information of the geocache's creator</p>"          },          {            "group": "Success 200",            "type": "Number",            "optional": false,            "field": "xCoord",            "description": "<p>X-coordinate of the geocache</p>"          },          {            "group": "Success 200",            "type": "Number",            "optional": false,            "field": "yCoord",            "description": "<p>Y-coordinate of the geocache</p>"          },          {            "group": "Success 200",            "type": "Log[]",            "optional": false,            "field": "logs",            "description": "<p>Array of log information</p>"          },          {            "group": "Success 200",            "type": "Date",            "optional": false,            "field": "updated",            "description": "<p>Date geocache was last updated</p>"          }        ]      },      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n[\n {\n     \"_id\": \"57314883c1d0de878fb4f137\",\n     \"yCoord\": 157.312452463,\n     \"xCoord\": -131.31785312,\n     \"createdBy\": {\n         \"id\": \"573147edc1d0de878fb4f136\",\n         \"username\": \"JohnDoe\",\n         \"email\": \"john@examplemail.com\",\n         \"memberSince\": \"2016-05-10T02:31:09.692Z\",\n         \"geocachesVisited\": [],\n         \"geocachesCreated\": [\n             \"57314883c1d0de878fb4f137\"\n         ]\n     },\n     \"description\": \"Test Description\",\n     \"name\": \"John's Cache\",\n     \"__v\": 0,\n     \"logs\": [],\n     \"updated\": \"2016-05-10T02:33:39.129Z\"\n },\n {\n     \"_id\": \"57314883c1d0de878fb4f138\",\n     \"yCoord\": 157.234452463,\n     \"xCoord\": -130.317321412,\n     \"createdBy\": {\n         \"id\": \"573147edc1d0de878fb4f135\",\n         \"username\": \"JaneDoe\",\n         \"email\": \"jane@examplemail.com\",\n         \"memberSince\": \"2015-12-11T02:31:09.692Z\",\n         \"geocachesVisited\": [],\n         \"geocachesCreated\": [\n             \"57314883c1d0de878fb4f138\"\n         ]\n     },\n     \"description\": \"Test Description 2\",\n     \"name\": \"Jane's Cache\",\n     \"__v\": 0,\n     \"logs\": [],\n     \"updated\": \"2016-05-10T02:33:39.129Z\"\n }\n]",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "routes/geocaches.js",    "groupTitle": "Geocaches"  },  {    "type": "put",    "url": "api/geocaches/{geocache_id}",    "title": "Updates a geocache",    "name": "UpdateGeocache",    "group": "Geocaches",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "name",            "description": "<p>New name of the geocache</p>"          },          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "description",            "description": "<p>New description of the geocache</p>"          },          {            "group": "Parameter",            "type": "Number",            "optional": false,            "field": "xCoord",            "description": "<p>New X-coordinate of the geocache</p>"          },          {            "group": "Parameter",            "type": "Number",            "optional": false,            "field": "yCoord",            "description": "<p>New Y-coordinate of the geocache</p>"          },          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "token",            "description": "<p>Authentication token of existing user (see api/authentication).</p>"          }        ]      }    },    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"message\": \"Geocache has been updated\",\n  \"user\": {\n    \"_id\": \"57314883c1d0de878fb4f137\",\n    \"yCoord\": 107.312452463,\n    \"xCoord\": -121.31785312,\n    \"createdBy\": {\n        \"id\": \"573147edc1d0de878fb4f136\",\n        \"username\": \"JohnDoe\",\n        \"email\": \"john@examplemail.com\",\n        \"memberSince\": \"2016-05-10T02:31:09.692Z\",\n        \"geocachesVisited\": [],\n        \"geocachesCreated\": [\n            \"57314883c1d0de878fb4f137\"\n        ]\n    },\n    \"description\": \"Trial Description\",\n    \"name\": \"John's Better Cache\",\n    \"__v\": 0,\n    \"logs\": [],\n    \"updated\": \"2016-05-10T02:33:39.129Z\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "routes/geocaches.js",    "groupTitle": "Geocaches"  },  {    "type": "post",    "url": "api/users",    "title": "Creates a new user",    "name": "CreateUser",    "group": "Users",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "username",            "description": "<p>Username of the new user</p>"          },          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "password",            "description": "<p>Password of the new user</p>"          },          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "email",            "description": "<p>Email of the new user</p>"          },          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "token",            "description": "<p>Authentication token of existing user (see api/authentication).</p>"          }        ]      }    },    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"message\": \"User was created\",\n  \"user\": {\n     \"_id\": \"1q2w3e4r56t7yu8\"\n     \"username\": \"JohnDoe\",\n     \"email\": \"john@examplemail.com\"\n     \"geocachesVisited\": \"[1234wer45rty73e2d, q2w3e4df5t6y879u]\"\n     \"geocachesCreated\": \"[q2w3e4df5t6y879u]\"\n  }\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "routes/users.js",    "groupTitle": "Users"  },  {    "type": "delete",    "url": "api/users/{user_id}",    "title": "Deletes a user by id",    "name": "DeleteUser",    "group": "Users",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "token",            "description": "<p>Authentication token of existing user (see api/authentication).</p>"          }        ]      }    },    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"message\": \"User successfully deleted\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "routes/users.js",    "groupTitle": "Users"  },  {    "type": "get",    "url": "api/users/{user_id}",    "title": "Requests user information",    "name": "GetUserFromId",    "group": "Users",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "token",            "description": "<p>Authentication token of existing user (see api/authentication).</p>"          }        ]      }    },    "success": {      "fields": {        "Success 200": [          {            "group": "Success 200",            "type": "String",            "optional": false,            "field": "_id",            "description": "<p>Unique id of the user</p>"          },          {            "group": "Success 200",            "type": "String",            "optional": false,            "field": "username",            "description": "<p>Username of the user</p>"          },          {            "group": "Success 200",            "type": "String",            "optional": false,            "field": "email",            "description": "<p>Email of the user</p>"          },          {            "group": "Success 200",            "type": "Date",            "optional": false,            "field": "memberSince",            "description": "<p>Date when user joined</p>"          },          {            "group": "Success 200",            "type": "Geocache[]",            "optional": false,            "field": "geocachesVisited",            "description": "<p>Geocaches that the user has logged visits</p>"          },          {            "group": "Success 200",            "type": "Geocache[]",            "optional": false,            "field": "geocachesCreated",            "description": "<p>Geocaches that the user has created</p>"          }        ]      },      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n     \"_id\": \"1q2w3e4r56t7yu8\"\n     \"username\": \"JohnDoe\",\n     \"email\": \"john@examplemail.com\"\n     \"memberSince\": \"2016-05-10T02:31:09.692Z\"\n     \"geocachesVisited\": \"[]\"\n     \"geocachesCreated\": \"[]\"\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "routes/users.js",    "groupTitle": "Users"  },  {    "type": "get",    "url": "api/users",    "title": "Requests all user information",    "name": "GetUsers",    "group": "Users",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "token",            "description": "<p>Authentication token of existing user (see api/authentication).</p>"          }        ]      }    },    "success": {      "fields": {        "Success 200": [          {            "group": "Success 200",            "type": "String",            "optional": false,            "field": "_id",            "description": "<p>Unique id of the user</p>"          },          {            "group": "Success 200",            "type": "String",            "optional": false,            "field": "username",            "description": "<p>Username of the user</p>"          },          {            "group": "Success 200",            "type": "String",            "optional": false,            "field": "email",            "description": "<p>Email of the user</p>"          },          {            "group": "Success 200",            "type": "Date",            "optional": false,            "field": "memberSince",            "description": "<p>Date when user joined</p>"          },          {            "group": "Success 200",            "type": "Geocache[]",            "optional": false,            "field": "geocachesVisited",            "description": "<p>Geocaches that the user has logged visits</p>"          },          {            "group": "Success 200",            "type": "Geocache[]",            "optional": false,            "field": "geocachesCreated",            "description": "<p>Geocaches that the user has created</p>"          }        ]      },      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n[\n {\n     \"_id\": \"1q2w3e4r56t7yu8\"\n     \"username\": \"JohnDoe\",\n     \"email\": \"john@examplemail.com\"\n     \"memberSince\": \"2016-05-10T02:31:09.692Z\"\n     \"geocachesVisited\": \"[]\"\n     \"geocachesCreated\": \"[]\"\n },\n {\n     \"_id\": \"9u89u7t8eyqwqwdq0\"\n     \"username\": \"JaneDoe\",\n     \"email\": \"jane@examplemail.com\"\n     \"memberSince\": \"\"2016-05-10T01:31:19.692Z\"\"\n     \"geocachesVisited\": \"[]\"\n     \"geocachesCreated\": \"[]\"\n }\n]",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "routes/users.js",    "groupTitle": "Users"  },  {    "type": "put",    "url": "api/users/{user_id}",    "title": "Updates a user",    "name": "UpdateUser",    "group": "Users",    "parameter": {      "fields": {        "Parameter": [          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "username",            "description": "<p>New username of the user</p>"          },          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "email",            "description": "<p>New email of the user</p>"          },          {            "group": "Parameter",            "type": "String",            "optional": false,            "field": "token",            "description": "<p>Authentication token of existing user (see api/authentication).</p>"          }        ]      }    },    "success": {      "examples": [        {          "title": "Success-Response:",          "content": "HTTP/1.1 200 OK\n{\n  \"message\": \"User has been updated\",\n  \"user\": {\n     \"_id\": \"1q2w3e4r56t7yu8\"\n     \"username\": \"JohnDoel\",\n     \"email\": \"johnnyboy@examplemail.com\"\n     \"geocachesVisited\": \"[1234wer45rty73e2d, q2w3e4df5t6y879u]\"\n     \"geocachesCreated\": \"[q2w3e4df5t6y879u]\"\n  }\n}",          "type": "json"        }      ]    },    "version": "0.0.0",    "filename": "routes/users.js",    "groupTitle": "Users"  }] });
