+++ api/docs/api_project.js
@@ -1,1 +1,1 @@
-define({  "name": "Geothing API",  "version": "0.0.1",  "description": "A geocaching API for the 21.16th century.",  "sampleUrl": false,  "apidoc": "0.2.0",  "generator": {    "name": "apidoc",    "time": "2016-05-11T02:04:03.388Z",    "url": "http://apidocjs.com",    "version": "0.16.1"  }});
+define({  "name": "Geothing API",  "version": "0.0.1",  "description": "A geocaching API for the 21.16th century.",  "sampleUrl": false,  "apidoc": "0.2.0",  "generator": {    "name": "apidoc",    "time": "2016-05-15T18:25:30.481Z",    "url": "http://apidocjs.com",    "version": "0.16.1"  }});
